+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 IMPROVEMENTS:
 
+- *New Resource*: _alicloud_log_machine_group_ ([#183](https://github.com/terraform-providers/terraform-provider-alicloud/pull/183))
 - *New Resource*: _alicloud_log_store_index_ ([#182](https://github.com/terraform-providers/terraform-provider-alicloud/pull/182))
 - *New Resource*: _alicloud_log_store_ ([#181](https://github.com/terraform-providers/terraform-provider-alicloud/pull/181))
 - *New Resource*: _alicloud_log_project_ ([#180](https://github.com/terraform-providers/terraform-provider-alicloud/pull/180))